# src/schemaOpenAPI.yaml
openapi: 3.1.0
info:
  title: Fatura Personalizada API
  description: API para gerar Faturas Personalizadas integrado a OS do Omie
  version: 0.0.15
  contact:
    name: Fabio
    email: fabio@oondemand.com.br
servers:
  - url: https://api.oondemand.com.br/fatura-personalizada
    description: Servidor de Produção
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Usuario:
      type: object
      properties:
        nome:
          type: string
          description: Nome do usuário
        email:
          type: string
          description: Email do usuário
        senha:
          type: string
          description: Senha do usuário
        status:
          type: string
          enum: [ativo, inativo, arquivado]
          description: Status do usuário
        permissoes:
          type: array
          items:
            type: string
          description: Permissões do usuário
      required:
        - nome
        - email
        - senha
    Configuracao:
      type: object
      properties:
        nome:
          type: string
        codigo:
          type: string
        valor:
          type: string
        baseOmie:
          type: string
...
    Log:
      type: object
      properties:
        usuario:
          type: string
          description: ID do usuário que fez a requisição
        endpoint:
          type: string
        metodo:
          type: string
        ip:
          type: string
        dadosRequisicao:
          type: object
        dadosResposta:
          type: object
        statusResposta:
          type: integer
paths:
...
  /auth/login:
    post:
      summary: Login de usuário
      operationId: loginUsuario
      requestBody:
        description: Dados de login
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                senha:
                  type: string
              required:
                - email
                - senha
      responses:
        '200':
          description: Token JWT gerado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Credenciais inválidas
  /usuarios:
    get:
      summary: Listar todos os usuários
      operationId: listarUsuarios
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      summary: Registrar um novo usuário
      operationId: registrarUsuario
      security:
        - BearerAuth: []
      requestBody:
        description: Dados para registrar um novo usuário
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '201':
          description: Usuário registrado com sucesso
  /usuarios/{id}:
    get:
      summary: Obter um usuário pelo ID
      operationId: obterUsuario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '404':
          description: Usuário não encontrado
    put:
      summary: Atualizar um usuário pelo ID
      operationId: atualizarUsuario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Dados atualizados do usuário
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '200':
          description: Usuário atualizado com sucesso
        '404':
          description: Usuário não encontrado
  /configuracoes:
    get:
      summary: Listar todas as configurações
      operationId: listarConfiguracoes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de configurações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Configuracao'
    post:
      summary: Criar uma nova configuração
      operationId: criarConfiguracao
      security:
        - BearerAuth: []
      requestBody:
        description: Dados para criar uma nova configuração
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuracao'
      responses:
        '201':
          description: Configuração criada com sucesso
  /configuracoes/{id}:
    get:
      summary: Obter uma configuração pelo ID
      operationId: obterConfiguracao
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuração encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuracao'
        '404':
          description: Configuração não encontrada
    patch:
      summary: Atualizar uma configuração pelo ID
      operationId: atualizarConfiguracao
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Dados atualizados da configuração
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuracao'
      responses:
        '200':
          description: Configuração atualizada com sucesso
        '404':
          description: Configuração não encontrada
    delete:
      summary: Deletar uma configuração pelo ID
      operationId: deletarConfiguracao
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Configuração deletada com sucesso
        '404':
          description: Configuração não encontrada
...
  /logs/rastreabilidade:
    get:
      summary: Obter logs de rastreabilidade
      operationId: getRastreabilidadeLogs
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
          description: Número da página
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: Número de registros por página
      responses:
        '200':
          description: Lista de logs de rastreabilidade
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalLogs:
                    type: integer
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Log'
        '500':
          description: Erro ao buscar logs